refactor: extract OpenAI message converter from SessionManager
Move buildOpenAIMessages, sessionMessageToOpenAIMessage, pairToolMessages,
getTrailingPendingToolCallMessage, and related helpers (~240 lines) into a
dedicated OpenAIMessageConverter class under src/common/.

SessionManager now delegates to the converter while retaining a deprecated
buildOpenAIMessages pass-through for test compatibility.

Net reduction: 2884 → 2644 lines in session.ts
